refactored schedoscope config into its own submodule
