+++ schedoscope-conf/src/main/resources/reference.conf
