+++ schedoscope-core/pom.xml
@@ -274,6 +274,11 @@
 		</dependency>
 		<dependency>
 			<groupId>schedoscope</groupId>
+			<artifactId>schedoscope-conf</artifactId>
+			<version>0.4.4-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>schedoscope</groupId>
 			<artifactId>schedoscope-export</artifactId>
 			<version>0.4.4-SNAPSHOT</version>
 			<classifier>jar-with-dependencies</classifier>
