+++ schedoscope-core/src/main/scala/org/schedoscope/scheduler/actors/TransformationManagerActor.scala
@@ -18,7 +18,7 @@ package org.schedoscope.scheduler.actors
 import akka.actor.SupervisorStrategy.{ Escalate, Restart }
 import akka.actor.{ Actor, ActorRef, OneForOneStrategy, Props, actorRef2Scala }
 import akka.event.{ Logging, LoggingReceive }
-import org.schedoscope.SchedoscopeSettings
+import org.schedoscope.conf.SchedoscopeSettings
 import org.schedoscope.dsl.View
 import org.schedoscope.dsl.transformations.{ FilesystemTransformation, Transformation }
 import org.schedoscope.scheduler.driver.{ Driver, RetryableDriverException }
