+++ schedoscope-metascope/src/main/java/org/schedoscope/metascope/conf/TestSpringConfiguration.java
@@ -17,7 +17,7 @@ package org.schedoscope.metascope.conf;
 
 import javax.sql.DataSource;
 
-import org.schedoscope.Schedoscope;
+import org.schedoscope.conf.BaseSettings;
 import org.schedoscope.metascope.index.SolrFacade;
 import org.schedoscope.metascope.tasks.repository.RepositoryDAO;
 import org.schedoscope.metascope.tasks.repository.jdbc.NativeSqlRepository;
@@ -36,6 +36,8 @@ import org.springframework.security.config.annotation.web.configuration.WebSecur
 import org.springframework.security.core.session.SessionRegistry;
 import org.springframework.security.core.session.SessionRegistryImpl;
 import org.springframework.security.web.session.HttpSessionEventPublisher;
+
+import com.typesafe.config.ConfigFactory;
 
 /**
  * Configuration to secure the application (login)
@@ -66,7 +68,7 @@ public class TestSpringConfiguration extends WebSecurityConfigurerAdapter {
 
   @Bean
   public MetascopeConfig metascopeConfig() {
-    return new MetascopeConfig(Schedoscope.settings());
+    return new MetascopeConfig(new BaseSettings(ConfigFactory.load()));
   }
 
   @Bean
